+++ src/assets/fonts/GilroyBold.ttf
