+++ src/assets/media/index.ts
@@ -1,40 +1,44 @@
 import error from 'assets/media/img/error.png';
+import alert from 'assets/media/svg/alert.svg';
 import arrow from 'assets/media/svg/arrow.svg';
 import avatar from 'assets/media/svg/avatar.svg';
-import avgMargin from 'assets/media/svg/avg-margin.svg';
-import avgValue from 'assets/media/svg/avg-value.svg';
-import chevronDown from 'assets/media/svg/chevron-down.svg';
-import chevronLeft from 'assets/media/svg/chevron-left.svg';
+import avgMargin from 'assets/media/svg/avgMargin.svg';
+import avgValue from 'assets/media/svg/avgValue.svg';
+import chevronDown from 'assets/media/svg/chevronDown.svg';
+import chevronLeft from 'assets/media/svg/chevronLeft.svg';
 import clipboard from 'assets/media/svg/clipboard.svg';
 import close from 'assets/media/svg/close.svg';
 import coins from 'assets/media/svg/coins.svg';
-import deleteIcon from 'assets/media/svg/delete-icon.svg';
-import directCosts from 'assets/media/svg/direct-costs.svg';
+import deleteIcon from 'assets/media/svg/deleteIcon.svg';
+import directCosts from 'assets/media/svg/directCosts.svg';
 import document from 'assets/media/svg/document.svg';
 import dollar from 'assets/media/svg/dollar.svg';
 import download from 'assets/media/svg/download.svg';
-import editIcon from 'assets/media/svg/edit-icon.svg';
+import editIcon from 'assets/media/svg/editIcon.svg';
+import ellipsis from 'assets/media/svg/ellipsis.svg';
 import email from 'assets/media/svg/email.svg';
 import employees from 'assets/media/svg/employees.svg';
-import gradientBackground from 'assets/media/svg/gradient-background.svg';
+import gradientBackground from 'assets/media/svg/gradientBackground.svg';
 import hamburger from 'assets/media/svg/hamburger.svg';
 import home from 'assets/media/svg/home.svg';
-import hourlyPrice from 'assets/media/svg/hourly-price.svg';
+import hourlyPrice from 'assets/media/svg/hourlyPrice.svg';
 import invoice from 'assets/media/svg/invoice.svg';
-import leadClosing from 'assets/media/svg/lead-closing.svg';
+import leadClosing from 'assets/media/svg/leadClosing.svg';
 import logo from 'assets/media/svg/logo.svg';
 import margin from 'assets/media/svg/margin.svg';
 import plus from 'assets/media/svg/plus.svg';
-import projectsNumber from 'assets/media/svg/projects-number.svg';
+import projectsNumber from 'assets/media/svg/projectsNumber.svg';
 import revenue from 'assets/media/svg/revenue.svg';
 import search from 'assets/media/svg/search.svg';
-import teamSize from 'assets/media/svg/team-size.svg';
-import totalValue from 'assets/media/svg/total-value.svg';
+import teamSize from 'assets/media/svg/teamSize.svg';
+import totalValue from 'assets/media/svg/totalValue.svg';
 import trash from 'assets/media/svg/trash.svg';
 import velocity from 'assets/media/svg/velocity.svg';
-import weeksOver from 'assets/media/svg/weeks-over.svg';
+import weeksOver from 'assets/media/svg/weeksOver.svg';
 
 export {
+	error,
+	alert,
 	arrow,
 	avatar,
 	avgMargin,
@@ -50,9 +54,9 @@ export {
 	dollar,
 	download,
 	editIcon,
+	ellipsis,
 	email,
 	employees,
-	error,
 	gradientBackground,
 	hamburger,
 	home,
