+++ src/features/auth/LogoBanner.tsx
@@ -2,7 +2,7 @@ import { logo } from 'assets/media';
 
 const LogoBanner = () => {
 	return (
-		<div className='logo-banner flex h-full w-full items-center justify-center bg-[url("src/assets/media/svg/gradient-background.svg")] bg-cover bg-no-repeat'>
+		<div className='logo-banner flex h-full w-full items-center justify-center bg-[url("src/assets/media/svg/gradientBackground.svg")] bg-cover bg-no-repeat'>
 			<div
 				className='h-1/2 w-full bg-cover bg-center'
 				style={{
