+++ tailwind.config.js
@@ -62,21 +62,21 @@ export default {
 				'crimson-blaze': '#FF4D4F',
 			},
 			fontFamily: {
-				'inter-regular': ['Inter-Regular', 'sans-serif'],
-				'inter-medium': ['Inter-Medium', 'sans-serif'],
-				'inter-semi-bold': ['Inter-SemiBold', 'sans-serif'],
-				'inter-bold': ['Inter-Bold', 'sans-serif'],
-				'gilroy-regular': ['Gilroy-Regular', 'sans-serif'],
-				'gilroy-medium': ['Gilroy-Medium', 'sans-serif'],
-				'gilroy-semi-bold': ['Gilroy-SemiBold', 'sans-serif'],
-				'gilroy-bold': ['Gilroy-Bold', 'sans-serif'],
-				'helvetica-neue-regular': ['HelveticaNeue-Regular', 'sans-serif'],
-				'helvetica-neue-medium': ['HelveticaNeue-Medium', 'sans-serif'],
-				'helvetica-neue-bold': ['HelveticaNeue-Bold', 'sans-serif'],
-				'opensans-regular': ['OpenSans-Regular', 'sans-serif'],
-				'opensans-medium': ['OpenSans-Medium', 'sans-serif'],
-				'opensans-semi-bold': ['OpenSans-SemiBold', 'sans-serif'],
-				'opensans-bold': ['OpenSans-Bold', 'sans-serif'],
+				'inter-regular': ['InterRegular', 'sans-serif'],
+				'inter-medium': ['InterMedium', 'sans-serif'],
+				'inter-semi-bold': ['InterSemiBold', 'sans-serif'],
+				'inter-bold': ['InterBold', 'sans-serif'],
+				'gilroy-regular': ['GilroyRegular', 'sans-serif'],
+				'gilroy-medium': ['GilroyMedium', 'sans-serif'],
+				'gilroy-semi-bold': ['GilroySemiBold', 'sans-serif'],
+				'gilroy-bold': ['GilroyBold', 'sans-serif'],
+				'helvetica-neue-regular': ['HelveticaNeueRegular', 'sans-serif'],
+				'helvetica-neue-medium': ['HelveticaNeueMedium', 'sans-serif'],
+				'helvetica-neue-bold': ['HelveticaNeueBold', 'sans-serif'],
+				'opensans-regular': ['OpenSansRegular', 'sans-serif'],
+				'opensans-medium': ['OpenSansMedium', 'sans-serif'],
+				'opensans-semi-bold': ['OpenSansSemiBold', 'sans-serif'],
+				'opensans-bold': ['OpenSansBold', 'sans-serif'],
 			},
 		},
 	},
